backend_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
django-filter==23.3
drf-yasg==1.21.7
django-redis==5.4.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
# WebSocket 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
PyJWT==2.8.0
# Pillow==10.1.0  # 이미지 처리 시 주석 해제
# psycopg2-binary==2.9.7  # PostgreSQL 사용 시 주석 해제
